requirements.txt
alpaca==1.0.0
alpaca-py
alpaca_trade_api==2.3.0
holidays
colorama==0.4.6
discord.py==2.2.2
matplotlib==3.5.3
numpy==1.24.3
ta-lib
openai==0.26.2
pytest==7.2.2
pytz==2022.2.1
